requirements.txt
accelerate==0.19.0
aiohttp==3.8.4
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
datasets==2.12.0
diffusers==0.16.1
dill==0.3.6
docker-pycreds==0.4.0
filelock==3.12.0
fonttools==4.39.4
frozenlist==1.3.3
fsspec==2023.5.0
ftfy==6.1.1
gitdb==4.0.10
GitPython==3.1.31
huggingface-hub==0.14.1
hydra-core==1.3.2
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
kiwisolver==1.4.4
matplotlib==3.7.1
multidict==6.0.4
multiprocess==0.70.14
numpy==1.24.3
omegaconf==2.3.0
open-clip-torch==2.20.0
packaging==23.1
pandas==2.0.1
pathtools==0.1.2
Pillow==9.5.0
protobuf==3.20.3
psutil==5.9.5
pyarrow==12.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
regex==2023.5.5
requests==2.30.0
responses==0.18.0
safetensors==0.3.1
scipy==1.10.1
sentencepiece==0.1.99
sentry-sdk==1.23.1
setproctitle==1.3.2
six==1.16.0
smmap==5.0.0
timm==0.9.2
torch==1.12.1+cu113
torch-fidelity==0.3.0
torchmetrics==0.11.4
torchvision==0.13.1+cu113
tqdm==4.65.0
typing-extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
wandb==0.15.3
wcwidth==0.2.6
xxhash==3.2.0
yarl==1.9.2
zipp==3.15.0
